+++ STEP2_REQUIREMENTS.md
@@ -9,14 +9,16 @@
 3. onChange, onBlur에서의 검증 로직을 어떻게 추상화할수있을까
 4. 다중 인풋 에러 핸들링
 5. '숫자만 입력 가능합니다' 에러 메시지 제거
-    - 왜? error를 배열로 관리 -> 앞선 에러만 표기 -> 이미 앞에서 에러가 발생했을 때 뒤 입력에서의 숫자 외 입력은 에러 메시지가 안 뜸 -> 일관성 x
-    - 숫자 외 입력에 대해서는 또 예외적으로 처리해야 하는데 이게 필요한 비용인가? x 아니라고 생각함
-    - 추가 이슈) 숫자 길이 맞게 입력한 뒤 숫자 외 입력 -> 에러 상태로 표시됨 -> but 실제로는 숫자 외 입력이 필터링됐기 때문에 valid라고 판단해야 함 -> 이런 케이스도 처리 필요
+    - 왜? error를 배열로 관리 -> 앞선 에러만 표기 -> 이미 앞에서 에러가 발생했을 때 뒤 입력에서의 숫자 외 입력은 에러 메시지가 안 뜸 -> 일관성 x (미해결)
+    - 숫자 외 입력에 대해서는 또 예외적으로 처리해야 하는데 이게 필요한 비용인가? x 아니라고 생각함 (해결)
+    - 추가 이슈) 숫자 길이 맞게 입력한 뒤 숫자 외 입력 -> 에러 상태로 표시됨 -> but 실제로는 숫자 외 입력이 필터링됐기 때문에 valid라고 판단해야 함 -> 이런 케이스도 처리 필요 (해결)
+    - => 숫자 외 입력에 대한 예외처리 불필요해짐 (sanitize) + 이미 max length인 경우에는 숫자 외 입력 에러 안 뜸
 6. 정규식으로 value 필터링 vs 애초에 change에서 원천차단(early return)
     - 정규식으로 sanitize하는 방식이 더 나을 듯
     - 언제는 return하고 언제는 update하는 것보다 항상 sanitize된 값으로 update하는 방식이 더 흐름 읽기 좋음
     - 근데 이게 판단 기준이 되어도 되는 건지는 잘 모르겠음
 7. 어떤 경우에 numberOnly고, 어떤 경우에 required인지까지 그 검증 로직까지 하나로 묶어서 추출할 수 있을까
+    - type 속성을 기준으로 필터링
 8. M/Y 검증이 입력 즉시 필요할지 blur 시에 하면 될지
     - 현재 블러 검증: required, length -> 에러를 띄울 때 입력이 완료됨을 판단할 수 없음
     - M/Y 검증: 2자를 모두 채우면 입력이 완료됐다고 판단 가능 -> 바로 에러처리해주는 편이 자연스럽다고 생각
@@ -40,7 +42,14 @@
 5. input 포커싱 등 공통 로직을 훅으로 분리한다
    5-1. 이때 훅 안에서 고립된 상태를 만들지 않고, 부모로부터 전달받은 상태와 핸들러를 조작하는 순수 '기능 로직'의 역할만 하도록 구성한다.
     - 우선 setter 호출을 1회로 정리하는 것부터
-    -
+    - 함수 내부 if문 분기 방식 -> validates 구조 도입해 단순화
+
+   a. useErrorStatusList 훅과 useErrorStatus 훅을 구현?
+    - validates를 받아서 errorStatusList/errorStatus, handleChange, handleBlur를 반환하는 훅
+    - status state 관련 내부 로직을 캡슐화하면서 + 공통 로직 묶을 수 있음
+    - index 등 optional 파라미터 관련 타입 안정성 + 확장성 확보 필요함
+
+   b. 다른 방법 있나
 
 3. field별 validate를 기반으로 step을 관리한다
    3-1. 파생값으로써 step을 관리한다.
