+++ src/components/domain/ExpirationPeriodField.tsx
@@ -2,17 +2,15 @@ import { css } from '@emotion/react';
 import FormField, { type FormFieldProps } from '../ui/FormField';
 import Input from '../ui/Input';
 import type { CardInfo, Validate } from '../../types';
-import { useState } from 'react';
 import type { ExpirationPeriodErrorStatus } from '../../types';
 import { isNumber, isValidMonth, isValidYear, sanitizeNumber } from '../../utils';
 import { EXPIRATION_PERIOD_ERROR_MESSAGES, PERIOD_LENGTH_PER_INPUT } from '../../constants';
+import { useErrorStatusList } from '../../hooks/useErrorStatusList.ts';
 
 interface ExpirationPeriodFieldProps {
   value: CardInfo['expirationPeriod'];
   onUpdated: (value: CardInfo['expirationPeriod']) => void;
 }
-
-type ErrorStatusList = [ExpirationPeriodErrorStatus, ExpirationPeriodErrorStatus];
 
 const validates: Validate<ExpirationPeriodErrorStatus>[] = [
   {
@@ -45,17 +43,9 @@ const validates: Validate<ExpirationPeriodErrorStatus>[] = [
 ];
 
 export default function ExpirationPeriodField({ value, onUpdated }: ExpirationPeriodFieldProps) {
-  const [errorStatusList, setErrorStatusList] = useState<ErrorStatusList>([null, null]);
+  const { errorStatusList, onChange, onBlur } = useErrorStatusList(validates, PERIOD_LENGTH_PER_INPUT);
   const activeErrorStatus = errorStatusList.filter((errorStatus) => !!errorStatus)[0];
   const activeErrorIndex = errorStatusList.findIndex((errorStatus) => errorStatus === activeErrorStatus);
-
-  const setErrorStatus = (status: ExpirationPeriodErrorStatus, index: number) => {
-    setErrorStatusList((prev) => {
-      const updated = [...prev] as ErrorStatusList;
-      updated[index] = status;
-      return updated;
-    });
-  };
 
   const updateFormValue = (inputValue: string, index: number) => {
     const newValue = [...value] as CardInfo['expirationPeriod'];
@@ -64,25 +54,11 @@ export default function ExpirationPeriodField({ value, onUpdated }: ExpirationPe
   };
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>, index: number) => {
-    const inputValue = e.target.value;
-
-    const changeValidates = validates.filter((validate) => validate.type.includes('change'));
-    const activeValidate = changeValidates.find((validate) => validate.rule(inputValue, index));
-
-    setErrorStatus(activeValidate?.errorStatus ?? null, index);
-    updateFormValue(inputValue, index);
+    onChange(e, index);
+    updateFormValue(e.target.value, index);
   };
 
-  const handleBlur = (e: React.FocusEvent<HTMLInputElement>, index: number) => {
-    const inputValue = e.target.value;
-
-    const blurValidates = validates.filter((validate) => validate.type.includes('blur'));
-    const activeValidate = blurValidates.find((validate) => validate.rule(inputValue));
-
-    if (activeValidate) {
-      setErrorStatus(activeValidate.errorStatus, index);
-    }
-  };
+  const handleBlur = onBlur;
 
   const formFieldProps: Omit<FormFieldProps, 'children'> = {
     title: '카드 유효기간을 입력해 주세요',
