+++ src/types.ts
@@ -10,7 +10,7 @@ export type ErrorStatus = null | 'required' | 'invalidLength' | 'numberOnly';
 
 export type ExpirationPeriodErrorStatus = ErrorStatus | 'invalidMonth' | 'invalidYear';
 
-export interface Validate<T = ErrorStatus> {
+export interface Validate<T> {
   type: ('change' | 'blur')[];
   rule: (inputValue: string, index?: number) => boolean;
   errorStatus: T;
